delete plan from directory collection'
