+++ portfolioplanning.code-workspace
@@ -31,6 +31,7 @@
 			//"**/*config.js": true,
         },
         "prettier.tabWidth": 4,
-        "prettier.printWidth": 120
+        "prettier.printWidth": 120,
+        "prettier.requireConfig": true
 	}
 }
